+++ R/utils.R
@@ -141,7 +141,7 @@ any_of2 = function(x, ignore.case=TRUE, ...){
 #' @keywords internal
 #' @importFrom cli cli_warn
 #' @importFrom dplyr select
-get_data_name = function(df, crfname=getOption("edc_crfname", "crfname")){
+get_data_name = function(df, crfname=getOption("edc_cols_crfname", "crfname")){
   sel = select(df, any_of2(crfname))
   if(!is.null(attr(df, "data_name"))){
     attr(df, "data_name")