+++ tests/testthat/test-utils.R
@@ -135,6 +135,13 @@ test_that("7zip not in the path", {
 
 
 
+# Options -------------------------------------------------------------------------------------
+
+
+test_that("No missing options", {
+  missing_options = missing_options_helper()
+  expect_identical(missing_options, character(0))
+})
 
 # Expect --------------------------------------------------------------------------------------